오늘의 프로젝트 작업(헤더, css 등등 수정)

+++ kim/프로젝트 작업용/asia/src/main/java/com/asia/repository/SeatARepository.java
@@ -18,7 +18,7 @@ public interface SeatARepository extends JpaRepository<SeatA, Long> {
 	List<CountDto> getCount();
 	
 	@Query("select new com.asia.dto.SeatADto(num, stat, seat) from SeatA where application_num = :num")
-	List<SeatADto> getSeat(int num);
+	List<SeatADto> getSeat(Long num);
 
 	SeatA findByNum(Long num);
